# openapi.yaml
openapi: 3.1.0
info:
  title: NovaMember API
  version: 1.0.0
  description: |
    Shopify + Recharge membership management platform with AI-powered engagement scoring.
    All authenticated endpoints require a Bearer JWT obtained from `/api/auth/login` or `/api/checkout`.
  contact:
    name: NovaMember Support
    email: support@novamember.io
  license:
    name: Proprietary
servers:
  - url: https://novamember.io
    description: Production
  - url: http://localhost:3000
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /api/auth/login:
    post:
      tags: [Authentication]
      summary: Log in with email + password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:    { type: string, format: email }
                password: { type: string, minLength: 1 }
      responses:
        '200':
          description: Authenticated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '422':
          description: Validation error
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '429':
          $ref: '#/components/responses/RateLimited'
...
  /api/auth/shopify/install:
    get:
      tags: [Authentication]
      summary: Start Shopify OAuth flow
      description: |
        Validates the shop domain, generates a CSRF nonce, stores it in the DB,
        and redirects to Shopify's OAuth consent screen.
      security: []
      parameters:
        - name: shop
          in: query
          required: true
          schema: { type: string, example: 'mystore.myshopify.com' }
      responses:
        '302':
          description: Redirect to Shopify OAuth
        '400':
          description: Invalid shop domain
  /api/auth/shopify/callback:
    get:
      tags: [Authentication]
      summary: Shopify OAuth callback
      description: |
        Validates HMAC and nonce, exchanges code for access token, registers webhooks,
        upserts the member, and redirects to the dashboard with a JWT.
      security: []
      parameters:
        - { name: code,      in: query, required: true,  schema: { type: string } }
        - { name: hmac,      in: query, required: true,  schema: { type: string } }
        - { name: shop,      in: query, required: true,  schema: { type: string } }
        - { name: state,     in: query, required: true,  schema: { type: string } }
        - { name: timestamp, in: query, required: true,  schema: { type: string } }
      responses:
        '302':
          description: Redirect to /dashboard?token=...
        '400':
          description: HMAC invalid or nonce expired

  # ── Checkout ──────────────────────────────────────────────────────────────────
  /api/checkout:
    post:
      tags: [Checkout]
      summary: Create a paid subscription
      description: |
        Full checkout flow:
        1. Create Shopify customer
        2. Create Recharge customer + subscription (idempotent)
        3. Create NovaMember account + subscription record (atomic transaction)
        4. Seed onboarding tasks + record CPRA consents
        5. Return signed JWT + subscription confirmation
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CheckoutBody' }
      responses:
        '201':
          description: Account created and subscription active
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CheckoutResponse' }
        '400':
          description: Unknown plan ID
        '409':
          description: Email already registered
        '422':
          description: Validation error
        '502':
          description: Payment processor (Recharge) error

  # ── Subscriptions ─────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 15-minute access JWT from `/api/auth/login` or `/api/checkout`
    cronSecret:
      type: apiKey
      in: header
      name: x-cron-secret
      description: Vercel cron secret from `CRON_SECRET` env var
    rechargeHmac:
      type: apiKey
      in: header
      name: X-Recharge-Hmac-Sha256
    shopifyHmac:
      type: apiKey
      in: header
      name: X-Shopify-Hmac-Sha256
  responses:
...
    RateLimited:
      description: Too many requests
      headers:
        Retry-After:
          schema: { type: integer }
          description: Seconds until the rate limit resets
        X-RateLimit-Limit:
          schema: { type: integer }
        X-RateLimit-Remaining:
          schema: { type: integer }
        X-RateLimit-Reset:
          schema: { type: integer }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      required: [error, code]
      properties:
        error:   { type: string }
        code:    { type: string, enum: [VALIDATION_ERROR, UNAUTHORIZED, FORBIDDEN, NOT_FOUND, CONFLICT, RATE_LIMITED, EXTERNAL_SERVICE_ERROR, PAYMENT_REQUIRED, INTERNAL_ERROR] }
        details: { type: object, nullable: true }
    AuthResponse:
      type: object
      properties:
        token:
          type: string
          description: Access JWT (15 min)
        refreshToken:
          type: string
          description: Refresh token (90 days)
        member:
          $ref: '#/components/schemas/MemberSummary'
    MemberSummary:
      type: object
      properties:
        id:        { type: string }
        email:     { type: string, format: email }
        firstName: { type: string }
        lastName:  { type: string }
        role:      { type: string, enum: [MEMBER, ADMIN] }
    CheckoutBody:
      type: object
      required: [firstName, lastName, email, password, planId, cpraConsent]
      properties:
        firstName:        { type: string, minLength: 1, maxLength: 100 }
        lastName:         { type: string, minLength: 1, maxLength: 100 }
        email:            { type: string, format: email }
        password:         { type: string, minLength: 8 }
        company:          { type: string, maxLength: 200 }
        planId:           { type: string, enum: [starter, growth, scale] }
        cpraConsent:      { type: boolean, enum: [true], description: 'Must be true' }
        marketingConsent: { type: boolean, default: false }
    CheckoutResponse:
      type: object
      properties:
        token:        { type: string }
        refreshToken: { type: string }
        member:       { $ref: '#/components/schemas/MemberSummary' }
        subscription: { $ref: '#/components/schemas/Subscription' }
        consent:
          type: object
          properties:
            dataProcessing: { type: boolean }
            marketing:      { type: boolean }
            recordedAt:     { type: string, format: date-time }
    Subscription:
      type: object
      properties:
        id:                     { type: string }
        plan:                   { type: string }
        price:                  { type: number }
        billingCycle:           { type: string }
        status:                 { type: string, enum: [ACTIVE, CANCELLED, PAUSED, PAST_DUE] }
        currentPeriodEnd:       { type: string, format: date-time }
        rechargeSubscriptionId: { type: string, nullable: true }
...